RagTool_IVghost_V2/requirements.txt
gradio
requests
PyPDF2
pandas
python-docx
openpyxl
fpdf
ollama
psutil
pdf2image
Pillow
pdfminer.six
python-magic